crawler/requirements.txt
flask==2.2.2
gunicorn==20.0.4
scrapy==2.7.1
beautifulsoup4==4.10.0
protego==0.1.16
python-slugify==6.1.0
python-dateutil==2.8.2